+++ packages/core/src/server/Server.ts
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2022 Javier Brea
+Copyright 2019-2023 Javier Brea
 Copyright 2019 XbyOrange
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
@@ -8,129 +8,90 @@ http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
+import http from "http";
+import type { Server as HttpServer } from "http";
 
-const http = require("http");
+import type { ConfigInterface, OptionInterface } from "@mocks-server/config";
+import type { LoggerInterface } from "@mocks-server/logger";
+import cors from "cors";
+import express from "express";
+import type { Application } from "express";
 
-const express = require("express");
-const cors = require("cors");
-const {
+import type { AlertsInterface } from "../alerts/Alerts.types";
+import { readFileSync } from "../common/helpers";
+
+import {
   addRequestId,
   jsonBodyParser,
   logRequest,
   urlEncodedBodyParser,
   notFound,
   errorHandler,
-} = require("./middlewares");
-const { readFileSync } = require("../common/helpers");
+} from "./middlewares";
+import type {
+  ServerConstructor,
+  ServerOptions,
+  ServerInterface,
+  Router,
+  CustomRouter,
+  ProtocolHttps,
+  ProtocolHttp,
+} from "./Server.types";
+import {
+  CORS_NAMESPACE,
+  JSON_BODY_PARSER_NAMESPACE,
+  URL_ENCODED_BODY_PARSER_NAMESPACE,
+  HTTPS_NAMESPACE,
+  OPTIONS,
+  CORS_OPTIONS,
+  JSON_BODY_PARSER_OPTIONS,
+  URL_ENCODED_BODY_PARSER_OPTIONS,
+  HTTPS_OPTIONS,
+  ALL_HOSTS,
+} from "./ServerOptions";
 
-const ALL_HOSTS = "0.0.0.0";
 const LOCALHOST = "localhost";
-
 const HTTPS_ALERT_ID = "https";
 const START_ALERT_ID = "start";
 const START_ERROR_MESSAGE = "Error starting server";
 const SERVER_ALERT_ID = "server";
 
-const OPTIONS = [
-  {
-    description: "Port number for the server to be listening at",
-    name: "port",
-    type: "number",
-    default: 3100,
-  },
-  {
-    description: "Host for the server",
-    name: "host",
-    type: "string",
-    default: ALL_HOSTS,
-  },
-];
+export const Server: ServerConstructor = class Server implements ServerInterface {
+  private _logger: LoggerInterface;
+  private _config: ConfigInterface;
+  private _mockRouter: Router;
+  private _customRouters: CustomRouter[];
+  private _alerts: AlertsInterface;
+  private _portOption: OptionInterface;
+  private _hostOption: OptionInterface;
+  private _corsEnabledOption: OptionInterface;
+  private _corsOptionsOption: OptionInterface;
+  private _jsonBodyParserEnabledOption: OptionInterface;
+  private _jsonBodyParserOptionsOption: OptionInterface;
+  private _urlEncodedBodyParserEnabledOption: OptionInterface;
+  private _urlEncodedBodyParserOptionsOption: OptionInterface;
+  private _httpsEnabledOption: OptionInterface;
+  private _httpsCertOption: OptionInterface;
+  private _httpsKeyOption: OptionInterface;
+  private _express: Application;
+  private _serverInitialized: boolean;
+  private _server?: HttpServer;
+  private _serverInitError?: Error;
+  private _serverStarted: boolean;
+  private _serverStartingPromise: Promise<void> | undefined;
+  private _serverStoppingPromise: Promise<void> | undefined;
 
-const HTTPS_NAMESPACE = "https";
-
-const HTTPS_OPTIONS = [
-  {
-    description: "Use https protocol or not",
-    name: "enabled",
-    type: "boolean",
-    default: false,
-  },
-  {
-    description: "Path to a TLS/SSL certificate",
-    name: "cert",
-    type: "string",
-  },
-  {
-    description: "Path to the certificate private key",
-    name: "key",
-    type: "string",
-  },
-];
-
-const CORS_NAMESPACE = "cors";
-
-const CORS_OPTIONS = [
-  {
-    description: "Use CORS middleware or not",
-    name: "enabled",
-    type: "boolean",
-    default: true,
-  },
-  {
-    description:
-      "Options for the CORS middleware. Further information at https://github.com/expressjs/cors#configuration-options",
-    name: "options",
-    type: "object",
-    default: {
-      preflightContinue: false,
-    },
-  },
-];
-
-const JSON_BODY_PARSER_NAMESPACE = "jsonBodyParser";
-
-const JSON_BODY_PARSER_OPTIONS = [
-  {
-    description: "Use json body-parser middleware or not",
-    name: "enabled",
-    type: "boolean",
-    default: true,
-  },
-  {
-    description:
-      "Options for the json body-parser middleware. Further information at https://github.com/expressjs/body-parser",
-    name: "options",
-    type: "object",
-    default: {},
-  },
-];
-
-const URL_ENCODED_BODY_PARSER_NAMESPACE = "urlEncodedBodyParser";
-
-const URL_ENCODED_BODY_PARSER_OPTIONS = [
-  {
-    description: "Use urlencoded body-parser middleware or not",
-    name: "enabled",
-    type: "boolean",
-    default: true,
-  },
-  {
-    description:
-      "Options for the urlencoded body-parser middleware. Further information at https://github.com/expressjs/body-parser",
-    name: "options",
-    type: "object",
-    default: { extended: true },
-  },
-];
-
-class Server {
-  static get id() {
+  static get id(): string {
     return "server";
   }
 
-  constructor({ config, alerts, routesRouter, logger }) {
+  constructor({ config, alerts, mockRouter, logger }: ServerOptions) {
     this._logger = logger;
     this._config = config;
+    this._mockRouter = mockRouter;
+    this._customRouters = [];
+    this._alerts = alerts;
+
     const corsConfigNamespace = this._config.addNamespace(CORS_NAMESPACE);
     const jsonBodyParserConfigNamespace = this._config.addNamespace(JSON_BODY_PARSER_NAMESPACE);
     const formBodyParserConfigNamespace = this._config.addNamespace(
@@ -153,7 +114,8 @@ class Server {
       httpsConfigNamespace.addOptions(HTTPS_OPTIONS);
 
     this.restart = this.restart.bind(this);
-    this._reinitServer = this._reinitServer.bind(this);
+    this._reinitializeServer = this._reinitializeServer.bind(this);
+    this._startServer = this._startServer.bind(this);
 
     this._hostOption.onChange(this.restart);
     this._portOption.onChange(this.restart);
@@ -163,18 +125,12 @@ class Server {
     this._jsonBodyParserOptionsOption.onChange(this.restart);
     this._urlEncodedBodyParserEnabledOption.onChange(this.restart);
     this._urlEncodedBodyParserOptionsOption.onChange(this.restart);
-    this._httpsEnabledOption.onChange(this._reinitServer);
-    this._httpsCertOption.onChange(this._reinitServer);
-    this._httpsKeyOption.onChange(this._reinitServer);
-
-    this._routesRouter = routesRouter;
-    this._customRouters = [];
-    this._alerts = alerts;
-
-    this._startServer = this._startServer.bind(this);
+    this._httpsEnabledOption.onChange(this._reinitializeServer);
+    this._httpsCertOption.onChange(this._reinitializeServer);
+    this._httpsKeyOption.onChange(this._reinitializeServer);
   }
 
-  init() {
+  public init(): Promise<void> {
     process.on("SIGINT", () => {
       this.stop().then(() => {
         this._logger.info("Server closed");
@@ -184,8 +140,8 @@ class Server {
     return Promise.resolve();
   }
 
-  _initServer() {
-    if (this._serverInitted) {
+  private async _initServer(): Promise<void> {
+    if (this._serverInitialized) {
       return;
     }
     this._logger.debug("Configuring server");
@@ -210,31 +166,32 @@ class Server {
     // TODO, move to variants router. Add options to routes to configure it
     this._express.use(logRequest({ logger: this._logger }));
     this._registerCustomRouters();
-    this._express.use(this._routesRouter);
+    this._express.use(this._mockRouter);
 
     // TODO, Add options to allow to disable or configure it
     this._express.use(notFound({ logger: this._logger }));
     this._express.use(errorHandler({ logger: this._logger }));
 
     // Create server
-    this._server = this._createServer();
+    this._server = await this._createServer();
     if (this._server) {
       this._alerts.remove(SERVER_ALERT_ID);
       this._server.on("error", (error) => {
         this._alerts.set(SERVER_ALERT_ID, "Server error", error);
         throw error;
       });
-      this._serverInitted = true;
+      this._serverInitialized = true;
     } else {
-      this._serverInitted = false;
+      this._alerts.set(SERVER_ALERT_ID, "Server not initialized");
+      this._serverInitialized = false;
     }
   }
 
-  _createHttpsServer() {
+  private async _createHttpsServer(): Promise<HttpServer | undefined> {
     this._logger.verbose("Creating HTTPS server");
     this._alerts.remove(HTTPS_ALERT_ID);
     try {
-      const https = require("https");
+      const https = await import("https");
       return https.createServer(
         {
           cert: readFileSync(this._httpsCertOption.value),
@@ -243,29 +200,29 @@ class Server {
         this._express
       );
     } catch (error) {
-      this._alerts.set(HTTPS_ALERT_ID, "Error creating HTTPS server", error);
-      this._serverInitError = error;
+      this._alerts.set(HTTPS_ALERT_ID, "Error creating HTTPS server", error as Error);
+      this._serverInitError = error as Error;
     }
   }
 
-  _createHttpServer() {
+  private _createHttpServer(): Promise<HttpServer> {
     this._logger.verbose("Creating HTTP server");
-    return http.createServer(this._express);
+    return Promise.resolve(http.createServer(this._express));
   }
 
-  _createServer() {
-    this._server = null;
-    return !!this._httpsEnabledOption.value ? this._createHttpsServer() : this._createHttpServer();
+  private _createServer(): Promise<HttpServer | undefined> {
+    this._server = undefined;
+    return this._httpsEnabledOption.value ? this._createHttpsServer() : this._createHttpServer();
   }
 
-  _reinitServer() {
-    if (this._serverInitted) {
-      this._serverInitted = false;
+  private _reinitializeServer(): Promise<void> {
+    if (this._serverInitialized) {
+      this._serverInitialized = false;
       if (this._serverStarted) {
         return this.restart();
       }
-      if (this._serverStarting) {
-        return this._serverStarting.then(() => {
+      if (this._serverStartingPromise) {
+        return this._serverStartingPromise.then(() => {
           return this.restart();
         });
       }
@@ -273,38 +230,34 @@ class Server {
     return Promise.resolve();
   }
 
-  _startServer(resolve, reject) {
+  private _startServer(resolve: () => unknown, reject: (error: Error) => unknown): void {
     const host = this._hostOption.value;
     const port = this._portOption.value;
 
     try {
-      this._server.listen(
+      const server = this._server as HttpServer;
+      server.listen(
         {
           port,
           host,
         },
-        (error) => {
-          if (error) {
-            this._serverStarting = false;
-            this._serverStarted = false;
-            this._alerts.set(START_ALERT_ID, START_ERROR_MESSAGE, error);
-            reject(error);
-          } else {
-            this._logger.info(`Server started and listening at ${this.url}`);
-            this._serverStarting = false;
-            this._serverStarted = true;
-            this._alerts.remove(START_ALERT_ID);
-            resolve(this);
-          }
+        () => {
+          this._alerts.remove(START_ALERT_ID);
+          this._logger.info(`Server started and listening at ${this.url}`);
+          this._serverStarted = true;
+          this._serverStartingPromise = undefined;
+          resolve();
         }
       );
     } catch (error) {
-      this._alerts.set(START_ALERT_ID, START_ERROR_MESSAGE, error);
-      reject(error);
+      this._alerts.set(START_ALERT_ID, START_ERROR_MESSAGE, error as Error);
+      this._serverStartingPromise = undefined;
+      this._serverStarted = false;
+      reject(error as Error);
     }
   }
 
-  _registerCustomRouters() {
+  private _registerCustomRouters(): void {
     this._logger.debug("Registering custom routers in server");
     this._customRouters.forEach((customRouter) => {
       this._logger.silly(`Registering custom router with path ${customRouter.path}`);
@@ -312,36 +265,43 @@ class Server {
     });
   }
 
-  stop() {
-    if (this._serverStopping) {
-      return this._serverStopping;
+  public stop(): Promise<void> {
+    if (this._serverStoppingPromise) {
+      return this._serverStoppingPromise;
     }
     if (this._server) {
-      this._serverStopping = new Promise((resolve) => {
+      const server = this._server as HttpServer;
+      this._serverStoppingPromise = new Promise((resolve) => {
         this._logger.verbose("Stopping server");
-        this._server.close(() => {
+        server.close(() => {
           this._logger.info("Server stopped");
           this._serverStarted = false;
-          this._serverStopping = false;
+          this._serverStoppingPromise = undefined;
           resolve();
         });
       });
-      return this._serverStopping;
+
+      return this._serverStoppingPromise;
     }
     return Promise.resolve();
   }
 
-  async start() {
-    this._initServer();
-    if (this._serverStarting) {
+  public async start(): Promise<void> {
+    await this._initServer();
+    if (this._serverStartingPromise) {
       this._logger.debug("Server is already starting, returning same promise");
-      return this._serverStarting;
+      return this._serverStartingPromise;
     }
-    this._serverStarting = new Promise(this._startServer);
-    return this._serverStarting;
+    this._serverStartingPromise = new Promise(this._startServer);
+    return this._serverStartingPromise;
   }
 
-  _getCustomRouterIndex(path, router) {
+  public async restart(): Promise<void> {
+    await this.stop();
+    return this.start();
+  }
+
+  private _getCustomRouterIndex(path: string, router: Router): number | null {
     let routerIndex = null;
     this._customRouters.forEach((customRouter, index) => {
       if (customRouter.path === path && customRouter.router === router) {
@@ -351,39 +311,32 @@ class Server {
     return routerIndex;
   }
 
-  addRouter(path, router) {
+  public addRouter(path: string, router: Router): Promise<void> {
     this._logger.info(`Adding custom router with path ${path}`);
     this._customRouters.push({
       path,
       router,
     });
-    return this._reinitServer();
+    return this._reinitializeServer();
   }
 
-  removeRouter(path, router) {
+  public removeRouter(path: string, router: Router): Promise<void> {
     this._logger.info(`Removing custom router with path ${path}`);
-    let indexToRemove = this._getCustomRouterIndex(path, router);
+    const indexToRemove = this._getCustomRouterIndex(path, router);
     if (indexToRemove !== null) {
       this._customRouters.splice(indexToRemove, 1);
-      return this._reinitServer();
+      return this._reinitializeServer();
     }
     return Promise.resolve();
   }
 
-  async restart() {
-    await this.stop();
-    return this.start();
+  public get protocol(): ProtocolHttp | ProtocolHttps {
+    return this._httpsEnabledOption.value ? "https" : "http";
   }
 
-  get protocol() {
-    return !!this._httpsEnabledOption.value ? "https" : "http";
-  }
-
-  get url() {
+  public get url(): string {
     const host = this._hostOption.value;
     const hostName = host === ALL_HOSTS ? LOCALHOST : host;
     return `${this.protocol}://${hostName}:${this._portOption.value}`;
   }
-}
-
-module.exports = Server;
+};